style_migrate/requirements.txt
absl-py==0.9.0
astor==0.8.1
gast==0.3.3
grpc-tools==0.0.1
grpcio==1.27.2
grpcio-tools==1.27.2
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.16.2
Pillow==7.0.0
protobuf==3.11.3
scipy==1.2.1
six==1.14.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
Werkzeug==1.0.0
